backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Cors
python-dotenv
psycopg2-binary
gunicorn
Werkzeug
flask_wtf
pytz